% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thyssenize.R
\name{display_progress}
\alias{display_progress}
\title{Display the progress if iterative processes}
\usage{
display_progress(n, nmax, t0, word)
}
\arguments{
\item{n}{Iteration step}

\item{nmax}{Number of iterations}

\item{t0}{initial time step}

\item{word}{Character string to define printed word}
}
\description{
Display the progress if iterative processes
}
\keyword{internal}
