% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thyssenize.R
\name{thyssen_i}
\alias{thyssen_i}
\title{Calculate area weighted values for regions based on station data for the time step i using Thyssen ploygons}
\usage{
thyssen_i(station_data, location, shape, bbox_sf, date_i)
}
\arguments{
\item{station_data}{Tibble providing the station time series data of a measured variable. The data must be organized in three columns where the first column provides the date, the second a unique station ID, and the third column the variable value. See examples for example table.}

\item{location}{sf object that only holds the corresponding station identifiers and the station location.}

\item{shape}{sf object providing polygons for which the area weighted means should be calculated.}

\item{bbox_sf}{bounding box sf object.}

\item{date_i}{Lubridate ymd format date defining the date for which the average values are calculated.}

\item{shape_id}{Polygon attribute that should be used as identifier for value averaging.}
}
\description{
Calculate area weighted values for regions based on station data for the time step i using Thyssen ploygons
}
